+++ index.js
@@ -2,7 +2,25 @@ const express = require('express');
 const app = express();
 const http = require('http').createServer(app);
 const io = require('socket.io')(http);
+const expressSession = require('express-session');
+const sharedSession = require('express-socket.io-session');
 const port = 8000;
+const session = expressSession({
+    secret: 'supah_secret_key',
+    resave: true,
+    saveUninitialized: true
+})
+
+const boundaries = {
+    top: 900,
+    left: 1500
+};
+
+app.use(session);
+
+io.use(sharedSession(session, {
+    autoSave: true
+}));
 
 app.get('/', function(req, res){
     res.sendFile(__dirname + '/index.html');
@@ -13,10 +31,11 @@ app.get('/public/style.css', function(req, res){
 });
 
 var playerId = 1;
+var players = {};
 var playerList = [];
 
 function createPlayer(id, socket){
-    playerList.push({
+    let player = {
         'id': id,
         'socket': socket,
         'directions': {
@@ -26,40 +45,46 @@ function createPlayer(id, socket){
             'left': false
         },
         'position': {
-            'left': 50,
-            'top': 50
-        }
-    })
+            'left': 250,
+            'top': 250
+        },
+        'active': true
+    }
+    players[id] = player;
+    playerList.push(id);
+    return player;
 }
 
 io.on('connection', function(socket){
-    socket.emit('assignId', playerId);
-    createPlayer(playerId, socket);
-    playerId++;
-    console.log('A user connected');
+    var session = socket.handshake.session;
+    if(session.playerId){
+        console.log('User id ' + session.playerId + ' connnected');
+        socket.emit('assignId', session.playerId);
+        let player = findPlayerObj(session.playerId);
+        player.active = true;
+        player.socket = socket;
+    } else {
+        socket.emit('assignId', playerId);
+        createPlayer(playerId, socket);
+        session.playerId = playerId;
+        session.save();
+        console.log('User id ' + playerId + ' connected');
+        playerId++;
+    }
+
     socket.on('disconnect', () => {
-        for(var i = 0; i < playerList.length; i++){
-            if(playerList[i].socket == socket){
-                playerList.splice(i, 1);
-            }
-        }
-        console.log('A user disconnected');
+        let id = session.playerId;
+        let player = players[id];
+        player.active = false;
+        console.log('User id ' + id + ' disconnected');
     });
     socket.on('startMove', (data) => {
-        let id = data.player;
-        for(var i = 0; i < playerList.length; i++){
-            if(playerList[i].id == id){
-                playerList[i].directions[data.direction] = true;
-            }
-        }
+        let player = players[data.playerId];
+        player.directions[data.direction] = true;
     });
     socket.on('endMove', (data) => {
-        let id = data.player;
-        for(var i = 0; i < playerList.length; i++){
-            if(playerList[i].id == id){
-                playerList[i].directions[data.direction] = false;
-            }
-        }
+        let player = players[data.playerId];
+        player.directions[data.direction] = false;
     })
 });
 
@@ -68,25 +93,29 @@ http.listen(port, () => {console.log("App now listening on port 8000")});
 function runGameLoop(){
     let playerPositions = [];
     for(var i = 0; i < playerList.length; i++){
-        let player = playerList[i];
-        if(player.directions.up){
-            player.position.top -= 10;
-        }
-        if (player.directions.down){
-            player.position.top += 10;
-        }
-        if(player.directions.right){
-            player.position.left += 10;
-        }
-        if (player.directions.left){
-            player.position.left -= 10;
+        let playerId = playerList[i];
+        let player = players[playerId];
+        if(!player.active) { continue; }
+        let directions = [null,'left','right','up','down'];
+        let positions = ['left', 'top'];
+        for(var i = 1; i < directions.length; i++){
+            let dir = directions[i];
+            if(player.directions[dir]){
+                let pos = positions[i / 2 > 1 ? 1 : 0];
+                i % 2 != 0 ? add = false : add = true
+                let pxSpeed = 5;
+                add ? newVal = player.position[pos] + pxSpeed : newVal = player.position[pos] - pxSpeed;
+                if(newVal > 0 && newVal < boundaries[pos]){
+                    player.position[pos] = newVal;
+                }
+            }
         }
         playerPositions.push({
-            id: player.id,
+            id: playerId,
             position: player.position
         });
     }
     io.emit('position', playerPositions);
 }
 
-setInterval(runGameLoop, 50);
+setInterval(runGameLoop, 15);