+++ CHANGELOG.md
@@ -1,9 +1,21 @@
 === master
 
+=== 0.0.4 (DaThresh)(2020-02-19)
+
+* Changed directional updates to a calculated function (DaThresh)(2020-02-19)
+* Added in express-session, socket.io-express-session hook (DaThresh)(2020-02-19)
+* Conserves player IDs based upon session (DaThresh)(2020-02-19)
+* Added 'active' key to player objects (DaThresh)(2020-02-19)
+* Added players as object hash (DaThresh)(2020-02-19)
+* Changed socket functions to utilize player hash system (DaThresh)(2020-02-19)
+* Changed playerList to be array of player IDs (DaThresh)(2020-02-19)
+* Added movement speed variable (DaThresh)(2020-02-19)
+* Implemented boundaries (DaThresh)(2020-02-19)
+
 === 0.0.3 (DaThresh)(2020-02-18)
 
 * Fixed removed player's rendering (DaThresh)(2020-02-18)
-* Changed server position emitter to emit once per game loop
+* Changed server position emitter to emit once per game loop (DaThresh)(2020-02-18)
 
 === 0.0.2 (DaThresh)(2020-02-18)